% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDatasets.r
\docType{data}
\name{deng}
\alias{deng}
\title{deng}
\usage{
data(deng)
}
\description{
Single-cell RNA-seq of mouse embryonic development from zygote to late blastocyst stage
}
\examples{
data(deng)
}
\keyword{datasets}
