% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDatasets.r
\name{exprs.Expression}
\alias{exprs.Expression}
\title{Extract the expression matrix}
\usage{
\method{exprs}{Expression}(gene)
}
\description{
Extract the expression matrix
}
\author{
Wuming Gong, \email{gongx030@umn.edu}
}
