% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDatasets.r
\docType{data}
\name{guo}
\alias{guo}
\title{guo}
\usage{
data(guo)
}
\description{
Resolution of Cell Fate Decisions Revealed by Single-Cell Gene Expression Analysis 
from Zygote to Blastocyst
}
\examples{
data(guo)
}
\keyword{datasets}
