% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDatasets.r
\docType{data}
\name{guo2}
\alias{guo2}
\title{guo2}
\usage{
data(guo2)
}
\description{
Single cell RNA-seq of the development of human primordial germ cells (PGC) and neighboring somatic cells from weeks 4 to 19 post-gestation
}
\examples{
data(guo2)
}
\keyword{datasets}
