% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDatasets.r
\docType{data}
\name{iPSC}
\alias{iPSC}
\title{iPSC}
\usage{
data(iPSC)
}
\description{
Single-cell RNA-seq of iPSC-CM differentiations at iPSC stage (D0), differentiation day 6 (D6),
D10, D30 and D60.
}
\examples{
data(iPSC)
}
\keyword{datasets}
