% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDatasets.r
\docType{data}
\name{loh}
\alias{loh}
\title{loh}
\usage{
data(loh)
}
\description{
Mapping the Pairwise Choices Leading from Pluripotency to Human Bone, Heart, and Other Mesoderm Cell Types
}
\examples{
data(loh)
}
\keyword{datasets}
