% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDatasets.r
\docType{data}
\name{petropoulos}
\alias{petropoulos}
\title{petropoulos}
\usage{
data(petropoulos)
}
\description{
Single-cell RNA-Seq of human preimplantation embryo development at E3, E4, E5, E6 and E7
}
\examples{
data(petropoulos)
}
\keyword{datasets}
