% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDatasets.r
\docType{data}
\name{pollen}
\alias{pollen}
\title{pollen}
\usage{
data(pollen)
}
\description{
Low-coverage single-cell mRNA sequencing reveals cellular heterogeneity 
and activated signaling pathways in developing cerebral cortex
}
\examples{
data(pollen)
}
\keyword{datasets}
