% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDatasets.r
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(X, min.expressed.gene = 2000, min.expressed.cell = 2,
  max.expressed.ratio = 1, normalize.by.size.effect = FALSE)
}
\arguments{
\item{X}{Gene expression matrix (Gene by Cell).}

\item{min.expressed.gene}{Cell level filtering criteria. For a given cell, if the number of expressed genes are less than min.expressed.gene, we filter it out.}

\item{min.expressed.cell}{Gene level filtering criteria. For a given gene, if the number of expressed cells are less than min.expressed.cell, we filter it out.}

\item{max.expressed.ratio}{Gene level filtering criteria. For a given gene, if the ratio of expressed cells are larger than max.expressed.ratio, we filter it out.}

\item{normalize.by.size.effect}{Normaize using size factor.}
}
\value{
Filtered gene expression matrix
}
\description{
A function for preprocessing gene expression matrix.
}
\examples{

library(scDatasets)
library(SummarizedExperiment)
data(usoskin)
X <- assays(usoskin)$count
X <- preprocess(X, min.expressed.gene = 0)

}
\references{
Wuming Gong, Il-Youp Kwak, Pruthvi Pota, Kaoko Koyano-Nakagawa and Daniel J. Garry (2017)
DrImpute: Imputing dropout eveents in single cell RNA sequencing data
}
\seealso{
\code{\link{DrImpute}}
}
\author{
Wuming Gong and Il-Youp Kwak
}
