% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDatasets.r
\name{preprocess2}
\alias{preprocess2}
\title{Process the scRNA-seq data matrix}
\usage{
preprocess2(X, min.expressed.gene = 2000, min.expressed.cell = 2,
  max.expressed.ratio = 0.9, normalize.by.size.effect = TRUE)
}
\description{
Process the scRNA-seq data matrix
}
