% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDatasets.r
\docType{package}
\name{scDatasets}
\alias{scDatasets}
\alias{scDatasets-package}
\title{scDatasets: Collections of processed single cell sequencing datasets}
\description{
Collections of processed single cell sequencing datasets
}
