% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDatasets.r
\docType{data}
\name{usoskin}
\alias{usoskin}
\alias{usoskin}
\title{usoskin}
\usage{
data(usoskin)

data(blakeley)
}
\description{
RNA-Seq of single cells from the mouse lumbar dorsal root ganglion

Single-Cell RNA-seq Defines the Three Cell Lineages of the Human Blastocyst
}
\examples{
data(usoskin)
data(blakeley)
}
\keyword{datasets}
